refactor(fe): extract GatewayStep and QrStep
